login: source session init in shell configs
If a shell is not started via login-inner (e.g. via ssh) it will not
source the sessionInit script. The solution is to source this file in
the respective config files for bash and zsh (more shell support can be
easily added).

The -u option is not necessary for the few lines in login-inner and
should not be activated after sourcing sessionInit in bash or zsh setup.