modules/build/activation: Prepend to PATH instead of setting it outright
Prior to this change: If your nix.conf has `max-jobs = 0` and `builders
= ssh://...`, thus configuring it to do all build jobs remotely,
activation would fail due to `nix-env` being unable to find `ssh` in
`PATH`. By prepending to `PATH` instead of overriding it entirely, this
is avoided.

Of course, `nix-env` probably shouldn't need a remote builder in order to
simply upgrade to an already-built derivation, but that's a Nix issue if
anything.